+++ CHANGELOG.md
@@ -1,1 +1,5 @@
-##### 03/23/2020
+##### 04/16/2020
+There is probably some official way to do these changelog files.
+
+Changes
+- change `edit` to `delete` on single address view navbar